requirements.txt
# Core ML/DL
tensorflow>=2.12.0
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
grad-cam>=1.4.8

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
pydantic>=2.0.0

# Experiment Tracking
mlflow>=2.8.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.6.0
pytest>=7.4.0
pytest-cov>=4.1.0

# Deployment
docker>=6.1.0
boto3>=1.29.0
